Fix spacing between status elements
Only the most right status element seemed to be padded with a space, for
the non-powerline configuration. The powerline configuration seems to
embed padding by the seperator, and thus requires a pad on the last
element.

This commit moves the padding next to the seperator to the script
output instead.

Remove trailing newline

Uniformalizes to the other scripts.