scripts: git: add option for untracked files
By default 'git status' will look for untracked files and the script will
always detect that the local tree has changes if untracked files are
present. That is not always optimal since sometimes we might have, for
example, some custom build scripts (for some project) that is always
present and we don't want the status bar to always indicate changes.

Hence, add an option to disable this behavior by adding the '-uno' flag
to 'git status'. The default behavior is still maintained.

Signed-off-by: Nuno Sá <noname.nuno@gmail.com>