git: add remote info support
With this new option, we get information about which remote branch we're
tracking. On top of this, we'll get information about ahead/behind
commits when we diverged from remote. The output format will be in the
form:

'local...remote +ahead -behind',

where ahead and behind are the number of commits ahead and behind.

This functionality is controlled by a new option called
'@dracula-git-show-remote-status'.

Note that for this to be properly displayed, we need to increase the
size of the right status bar when the git plugin is enabled.

In order to be easier to introduce the change, getMessage() was also a
bit changed in order to be easier to append the remote info.

Signed-off-by: Nuno Sá <noname.nuno@gmail.com>