allow user to force getting gpu info without asking lspci
i myself and other people have had issues with lspci not being available
or not returning the right values for data to be displayed, despite
there being an nvidia gpu and functioning nvidia-smi.
this commit adds an optional flag that allows to tell tmux that there is
an nvidia gpu, without any checks.