weather.sh: Introduce fallback when wttr.in reaches call limit
When wttr.in reaches call limit returns the following:

Sorry, we are running out of queries to the weather service at the moment.
Here is the weather report for the default city (just to show you what it looks like).
We will get new queries as soon as possible.
You can follow https://twitter.com/igor_chubin for the updates.

Which makes weather.sh display a random error.

Make the error predictable.