INSTALL.md: move back to root directory to fix Dracula website.
When the INSTALL.md directory was moved from the root of the repository,
it disappeared from the draculatheme.com website. Moving it back should
fix it.

It disappeared because the Dracula website relies on the contents of
INSTALL.md to generate the HTML for the page. When it was moved, it did
not know what to do.

Signed-off-by: Ethan Carter Edwards <ethan@ethancedwards.com>